refactor(products): 서비스 유즈케이스 별로 분리
