+++ src/main/java/kitchenpos/products/application/ChangeProductPriceService.java
@@ -3,10 +3,8 @@ package kitchenpos.products.application;
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.entity.Product;
 import kitchenpos.products.tobe.domain.repository.ProductRepository;
-import kitchenpos.products.tobe.domain.vo.ProductName;
 import kitchenpos.products.tobe.domain.vo.ProductPrice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,36 +15,24 @@ import java.util.NoSuchElementException;
 import java.util.UUID;
 
 @Service
-public class ProductService {
-    private final ProductRepository productRepository;
+public class ChangeProductPriceService {
     private final MenuRepository menuRepository;
-    private final PurgomalumClient purgomalumClient;
 
-    public ProductService(
-        final ProductRepository productRepository,
+    private final ProductRepository productRepository;
+
+    public ChangeProductPriceService(
         final MenuRepository menuRepository,
-        final PurgomalumClient purgomalumClient
+        final ProductRepository productRepository
     ) {
-        this.productRepository = productRepository;
         this.menuRepository = menuRepository;
-        this.purgomalumClient = purgomalumClient;
-    }
-
-    @Transactional
-    public Product create(final BigDecimal price, final String name) {
-        final ProductPrice productPrice = new ProductPrice(price);
-        final ProductName productName = new ProductName(name, purgomalumClient);
-
-        Product product = new Product(productName, productPrice);
-
-        return productRepository.save(product);
+        this.productRepository = productRepository;
     }
 
     @Transactional
     public Product changePrice(final UUID productId, final BigDecimal price) {
         final ProductPrice productPrice = new ProductPrice(price);
         final Product product = productRepository.findById(productId)
-            .orElseThrow(NoSuchElementException::new);
+                .orElseThrow(NoSuchElementException::new);
         product.changePrice(productPrice);
         final List<Menu> menus = menuRepository.findAllByProductId(productId);
         for (final Menu menu : menus) {
@@ -54,10 +40,10 @@ public class ProductService {
             for (final MenuProduct menuProduct : menu.getMenuProducts()) {
                 Product product1 = menuProduct.getProduct();
                 sum = sum.add(
-                    product1
-                        .getPrice()
-                            .getPrice()
-                            .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
+                        product1
+                                .getPrice()
+                                .getPrice()
+                                .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
                 );
             }
             if (menu.getPrice().compareTo(sum) > 0) {
@@ -65,10 +51,5 @@ public class ProductService {
             }
         }
         return product;
-    }
-
-    @Transactional(readOnly = true)
-    public List<Product> findAll() {
-        return productRepository.findAll();
     }
 }
