+++ src/test/java/kitchenpos/products/application/ChangeProductPriceServiceTest.java
@@ -1,0 +1,63 @@
+package kitchenpos.products.application;
+
+import kitchenpos.menus.application.InMemoryMenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.entity.Product;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullSource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+import static kitchenpos.Fixtures.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class ChangeProductPriceServiceTest {
+    private ProductRepository productRepository;
+    private MenuRepository menuRepository;
+    private ChangeProductPriceService changeProductPriceService;
+
+    @BeforeEach
+    void setUp() {
+        productRepository = new InMemoryProductRepository();
+        menuRepository = new InMemoryMenuRepository();
+        changeProductPriceService = new ChangeProductPriceService(menuRepository, productRepository);
+    }
+
+    @DisplayName("상품의 가격을 변경할 수 있다.")
+    @Test
+    void changePrice() {
+        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
+        BigDecimal price = BigDecimal.valueOf(15_000L);
+        final Product actual = changeProductPriceService.changePrice(productId, price);
+        assertThat(actual.getPrice().getPrice()).isEqualTo(price);
+    }
+
+    @DisplayName("상품의 가격이 올바르지 않으면 변경할 수 없다.")
+    @ValueSource(strings = "-1000")
+    @NullSource
+    @ParameterizedTest
+    void changePrice(final BigDecimal price) {
+        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
+        assertThatThrownBy(() -> changeProductPriceService.changePrice(productId, price))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("상품의 가격이 변경될 때 메뉴의 가격이 메뉴에 속한 상품 금액의 합보다 크면 메뉴가 숨겨진다.")
+    @Test
+    void changePriceInMenu() {
+        final Product product = productRepository.save(product("후라이드", 16_000L));
+        final Menu menu = menuRepository.save(menu(19_000L, true, menuProduct(product, 2L)));
+        BigDecimal price = BigDecimal.valueOf(8_000L);
+        changeProductPriceService.changePrice(product.getId(), price);
+        assertThat(menuRepository.findById(menu.getId()).get().isDisplayed()).isFalse();
+    }
+}
