--- src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -1,110 +1,0 @@
-package kitchenpos.products.application;
-
-import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.infra.PurgomalumClient;
-import kitchenpos.products.tobe.domain.entity.Product;
-import kitchenpos.products.tobe.domain.repository.ProductRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.UUID;
-
-import static kitchenpos.Fixtures.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-class ProductServiceTest {
-    private ProductRepository productRepository;
-    private MenuRepository menuRepository;
-    private PurgomalumClient purgomalumClient;
-    private ProductService productService;
-
-    @BeforeEach
-    void setUp() {
-        productRepository = new InMemoryProductRepository();
-        menuRepository = new InMemoryMenuRepository();
-        purgomalumClient = new FakePurgomalumClient();
-        productService = new ProductService(productRepository, menuRepository, purgomalumClient);
-    }
-
-    @DisplayName("상품을 등록할 수 있다.")
-    @Test
-    void create() {
-        BigDecimal price = BigDecimal.valueOf(16_000L);
-        String name = "후라이드";
-        final Product actual = productService.create(price, name);
-        assertThat(actual).isNotNull();
-        assertAll(
-            () -> assertThat(actual.getId()).isNotNull(),
-            () -> assertThat(actual.getName().getName()).isEqualTo(name),
-            () -> assertThat(actual.getPrice().getPrice()).isEqualTo(price)
-        );
-    }
-
-    @DisplayName("상품의 가격이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = "-1000")
-    @NullSource
-    @ParameterizedTest
-    void create(final BigDecimal price) {
-        String name = "후라이드";
-        assertThatThrownBy(() -> productService.create(price, name))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("상품의 이름이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = {"비속어", "욕설이 포함된 이름"})
-    @NullSource
-    @ParameterizedTest
-    void create(final String name) {
-        BigDecimal price = BigDecimal.valueOf(16_000L);
-        assertThatThrownBy(() -> productService.create(price, name))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("상품의 가격을 변경할 수 있다.")
-    @Test
-    void changePrice() {
-        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
-        BigDecimal price = BigDecimal.valueOf(15_000L);
-        final Product actual = productService.changePrice(productId, price);
-        assertThat(actual.getPrice().getPrice()).isEqualTo(price);
-    }
-
-    @DisplayName("상품의 가격이 올바르지 않으면 변경할 수 없다.")
-    @ValueSource(strings = "-1000")
-    @NullSource
-    @ParameterizedTest
-    void changePrice(final BigDecimal price) {
-        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
-        assertThatThrownBy(() -> productService.changePrice(productId, price))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("상품의 가격이 변경될 때 메뉴의 가격이 메뉴에 속한 상품 금액의 합보다 크면 메뉴가 숨겨진다.")
-    @Test
-    void changePriceInMenu() {
-        final Product product = productRepository.save(product("후라이드", 16_000L));
-        final Menu menu = menuRepository.save(menu(19_000L, true, menuProduct(product, 2L)));
-        BigDecimal price = BigDecimal.valueOf(8_000L);
-        productService.changePrice(product.getId(), price);
-        assertThat(menuRepository.findById(menu.getId()).get().isDisplayed()).isFalse();
-    }
-
-    @DisplayName("상품의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        productRepository.save(product("후라이드", 16_000L));
-        productRepository.save(product("양념치킨", 16_000L));
-        final List<Product> actual = productService.findAll();
-        assertThat(actual).hasSize(2);
-    }
-}
